@@ -5,7 +5,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=3.0.3 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,17 +15,18 @@ dependencies:
 
   # GetX
   get: ^4.6.5
-  get_storage: ^2.0.3
+  get_storage: ^2.1.1
 
   # Network
-  dio: ^4.0.6
+  dio: ^5.2.1+1
 
   # UI
-  shimmer: ^2.0.0
+  shimmer: ^3.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^2.0.0
 
 flutter:
   uses-material-design: true
